Move ScalaHtmlFormatter to report
